requirements.txt
django
djangorestframework
djoser
stream-chat
django-cors-headers
stream-python